refactor(swarm): extract client protocol contract to break the settlement-to-node cycle
The pseudosettle and swap settlement crates depended up on vertex-swarm-node
solely to import the ClientCommand/PseudosettleEvent/SwapEvent glue. Move the
shared command and event contract (plus FailureKind, the response-channel
aliases, ChunkTransferError, and RetrievalResult) down into the new
vertex-swarm-client-protocol crate, which both the node and the settlement
crates depend on. The node re-exports every item from its original paths, so
no downstream consumer changes.
