+++ crates/swarm/accounting/pseudosettle/Cargo.toml
@@ -10,7 +10,7 @@ description = "Time-based settlement provider for Swarm bandwidth accounting"
 
 [dependencies]
 vertex-swarm-accounting = { workspace = true }
-vertex-swarm-node = { workspace = true }
+vertex-swarm-client-protocol = { workspace = true }
 vertex-swarm-net-pseudosettle = { workspace = true }
 vertex-swarm-primitives = { workspace = true }
 vertex-swarm-api = { workspace = true }
