+++ crates/swarm/accounting/swap/Cargo.toml
@@ -14,7 +14,7 @@ workspace = true
 [dependencies]
 vertex-swarm-accounting-chequebook = { workspace = true }
 vertex-swarm-accounting = { workspace = true }
-vertex-swarm-node = { workspace = true, features = ["swap"] }
+vertex-swarm-client-protocol = { workspace = true, features = ["swap"] }
 vertex-swarm-primitives = { workspace = true }
 vertex-swarm-api = { workspace = true }
 vertex-tasks = { workspace = true }
