+++ crates/swarm/node/src/protocol/events.rs
@@ -1,372 +1,6 @@
-//! Events and commands for the client behaviour.
+//! Re-exports of the client command and event contract.
 //!
-//! The behaviour emits [`ClientEvent`]s and accepts [`ClientCommand`]s. Settlement
-//! events ([`PseudosettleEvent`], [`SwapEvent`]) are routed to their services via
-//! optional senders configured at construction time.
+//! The definitions live in `vertex-swarm-client-protocol` so the settlement
+//! services can share them without depending up on this crate.
 
-use alloy_primitives::U256;
-use libp2p::PeerId;
-use nectar_primitives::{AnyChunk, ChunkAddress};
-use tokio::sync::oneshot;
-use vertex_swarm_net_pseudosettle::PaymentAck;
-use vertex_swarm_net_pushsync::Receipt;
-#[cfg(feature = "swap")]
-use vertex_swarm_net_swap::SignedCheque;
-use vertex_swarm_primitives::{OverlayAddress, Stamp, StampedChunk, SwarmNodeType};
-
-use crate::client_service::{ChunkTransferError, RetrievalResult};
-
-/// Channel on which an outbound retrieval request resolves.
-///
-/// The sender travels with the request to the outbound substream; dropping it
-/// anywhere on that path surfaces as [`ChunkTransferError::Cancelled`].
-pub type RetrievalResponseTx = oneshot::Sender<Result<RetrievalResult, ChunkTransferError>>;
-
-/// Channel on which an outbound chunk push resolves.
-///
-/// A receipt whose storer cannot be recovered is rejected at decode and surfaces
-/// here as an error, never as a value.
-pub type PushResponseTx = oneshot::Sender<Result<Receipt, ChunkTransferError>>;
-
-/// Why a retrieval or pushsync request failed, classified for peer scoring.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FailureKind {
-    /// The peer delivered or pushed a chunk that failed address or stamp
-    /// reconstruction. Scored as invalid data.
-    InvalidChunk,
-    /// A transport, negotiation, timeout, or storer-reported failure that is
-    /// not evidence of malformed data. Scored as a plain failure.
-    Protocol,
-}
-
-/// Events emitted by the client behaviour.
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug, Clone)]
-pub enum ClientEvent {
-    /// Received a payment threshold from a peer.
-    PricingReceived {
-        /// The peer's overlay address.
-        peer: OverlayAddress,
-        /// The libp2p peer ID.
-        peer_id: PeerId,
-        /// The payment threshold announced by the peer.
-        threshold: U256,
-    },
-
-    /// Successfully sent our payment threshold to a peer.
-    PricingSent {
-        /// The peer we sent the threshold to.
-        peer: OverlayAddress,
-    },
-
-    /// We served an inbound retrieval request from our cache (scoring/metrics only).
-    InboundServed {
-        /// The peer we served.
-        peer: OverlayAddress,
-    },
-
-    /// We answered an inbound retrieval by forwarding to a closer peer.
-    InboundForwarded {
-        /// The peer we served.
-        peer: OverlayAddress,
-    },
-
-    /// We could not serve or forward an inbound retrieval; the substream reset.
-    InboundMissed {
-        /// The peer that asked.
-        peer: OverlayAddress,
-        /// The requested chunk address.
-        address: ChunkAddress,
-    },
-
-    /// We relayed a storer's receipt for an inbound pushsync (never signed it).
-    InboundRelayed {
-        /// The peer that pushed.
-        peer: OverlayAddress,
-    },
-
-    /// We stored an inbound pushsync delivery in the reserve and acknowledged it
-    /// with our own signed receipt. Reached only on a storer responsible for the
-    /// chunk (scoring/metrics only).
-    InboundStored {
-        /// The peer that pushed.
-        peer: OverlayAddress,
-    },
-
-    /// We could not forward an inbound pushsync (or, on the storer ingest path,
-    /// could not store or acknowledge it); the substream reset.
-    InboundPushFailed {
-        /// The peer that pushed.
-        peer: OverlayAddress,
-        /// The chunk address.
-        address: ChunkAddress,
-    },
-
-    /// Received a chunk from a peer (response to our request).
-    ChunkReceived {
-        /// The peer that sent the chunk.
-        peer: OverlayAddress,
-        /// The chunk address.
-        address: ChunkAddress,
-        /// The received chunk.
-        chunk: AnyChunk,
-        /// The postage stamp the responder attached, if any.
-        stamp: Option<Stamp>,
-        /// Time from request to delivery, for latency scoring.
-        latency: core::time::Duration,
-    },
-
-    /// A chunk retrieval request failed.
-    RetrievalFailed {
-        /// The peer we requested from.
-        peer: OverlayAddress,
-        /// The chunk address.
-        address: ChunkAddress,
-        /// Error description.
-        error: String,
-        /// Whether the failure was a malformed chunk (vs a plain failure).
-        kind: FailureKind,
-    },
-
-    /// Received a receipt for a chunk we pushed.
-    ReceiptReceived {
-        /// The peer that sent the receipt.
-        peer: OverlayAddress,
-        /// The chunk address.
-        address: ChunkAddress,
-        /// Time from push to receipt, for latency scoring.
-        latency: core::time::Duration,
-    },
-
-    /// A chunk push failed.
-    PushFailed {
-        /// The peer we pushed to.
-        peer: OverlayAddress,
-        /// The chunk address.
-        address: ChunkAddress,
-        /// Error description.
-        error: String,
-        /// Whether the failure was a malformed chunk (vs a plain failure).
-        kind: FailureKind,
-    },
-
-    /// A peer sent malformed data on an inbound substream; rejected at decode and
-    /// scored adversely for invalid data.
-    InboundInvalidData {
-        /// The peer that sent the malformed data.
-        peer: OverlayAddress,
-        /// The protocol that rejected the data.
-        protocol: &'static str,
-    },
-
-    /// The balance crossed the payment threshold; settle via swap or pseudosettle.
-    SettlementNeeded {
-        /// The peer to settle with.
-        peer: OverlayAddress,
-        /// Current balance (positive = they owe us).
-        balance: i64,
-    },
-
-    /// Received a pseudosettle payment from a peer.
-    PseudosettleReceived {
-        /// The peer that sent the payment.
-        peer: OverlayAddress,
-        /// The libp2p peer ID.
-        peer_id: PeerId,
-        /// The payment amount.
-        amount: U256,
-        /// Request ID for sending ack.
-        request_id: u64,
-    },
-
-    /// Successfully sent a pseudosettle payment.
-    PseudosettleSent {
-        /// The peer we sent to.
-        peer: OverlayAddress,
-        /// The libp2p peer ID.
-        peer_id: PeerId,
-        /// The ack received.
-        ack: PaymentAck,
-    },
-
-    /// Received a swap cheque from a peer.
-    #[cfg(feature = "swap")]
-    SwapChequeReceived {
-        /// The peer that sent the cheque.
-        peer: OverlayAddress,
-        /// The libp2p peer ID.
-        peer_id: PeerId,
-        /// The signed cheque received.
-        cheque: SignedCheque,
-        /// The peer's advertised exchange rate, from the headers exchange.
-        peer_rate: U256,
-    },
-
-    /// Successfully sent a swap cheque to a peer.
-    #[cfg(feature = "swap")]
-    SwapChequeSent {
-        /// The peer we sent the cheque to.
-        peer: OverlayAddress,
-        /// The libp2p peer ID.
-        peer_id: PeerId,
-        /// The peer's advertised exchange rate, from the headers exchange.
-        peer_rate: U256,
-    },
-
-    /// A peer's handler has been activated (after [`ClientCommand::ActivatePeer`]).
-    PeerActivated {
-        /// The libp2p peer ID.
-        peer_id: PeerId,
-        /// The peer's overlay address.
-        overlay: OverlayAddress,
-    },
-
-    /// A peer has disconnected.
-    PeerDisconnected {
-        /// The libp2p peer ID.
-        peer_id: PeerId,
-        /// The peer's overlay address.
-        overlay: OverlayAddress,
-    },
-
-    /// A protocol error occurred.
-    ProtocolError {
-        /// The peer involved (if known).
-        peer: Option<OverlayAddress>,
-        /// The libp2p peer ID (if known).
-        peer_id: Option<PeerId>,
-        /// The protocol that failed.
-        protocol: &'static str,
-        /// Error description.
-        error: String,
-    },
-}
-
-/// Commands accepted by the client behaviour.
-///
-/// Request commands ([`Self::RetrieveChunk`], [`Self::PushChunk`]) carry their
-/// response channel, so the enum is intentionally not `Clone`.
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug)]
-pub enum ClientCommand {
-    /// Activate the handler for a peer after handshake completes, transitioning it
-    /// from dormant to active.
-    ActivatePeer {
-        /// The libp2p peer ID.
-        peer_id: PeerId,
-        /// The peer's Swarm overlay address.
-        overlay: OverlayAddress,
-        /// The peer's node type.
-        node_type: SwarmNodeType,
-    },
-
-    /// Announce our payment threshold to a peer.
-    AnnouncePricing {
-        /// The peer to announce to.
-        peer: OverlayAddress,
-        /// The payment threshold to announce.
-        threshold: U256,
-    },
-
-    /// Request a chunk from a peer.
-    RetrieveChunk {
-        /// The peer to request from.
-        peer: OverlayAddress,
-        /// The chunk address to retrieve.
-        address: ChunkAddress,
-        /// Resolves with the retrieved chunk or the failure.
-        response: RetrievalResponseTx,
-    },
-
-    /// Push a chunk to a peer.
-    PushChunk {
-        /// The peer to push to.
-        peer: OverlayAddress,
-        /// The chunk address.
-        address: ChunkAddress,
-        /// The chunk and its postage stamp to push.
-        chunk: StampedChunk,
-        /// Resolves with the storer's receipt or the failure.
-        response: PushResponseTx,
-    },
-
-    /// Send a pseudosettle payment to a peer.
-    SendPseudosettle {
-        /// The peer to send the payment to.
-        peer: OverlayAddress,
-        /// The amount to pay.
-        amount: U256,
-    },
-
-    /// Acknowledge a pseudosettle payment.
-    AckPseudosettle {
-        /// The peer to ack.
-        peer: OverlayAddress,
-        /// Request ID from the received payment.
-        request_id: u64,
-        /// The ack to send.
-        ack: PaymentAck,
-    },
-
-    /// Send a swap cheque to a peer.
-    #[cfg(feature = "swap")]
-    SendCheque {
-        /// The peer to send the cheque to.
-        peer: OverlayAddress,
-        /// The signed cheque to send.
-        cheque: SignedCheque,
-    },
-
-    /// Disconnect from a peer.
-    ///
-    /// Used when a peer fails validation (e.g., threshold too low).
-    DisconnectPeer {
-        /// The peer to disconnect.
-        peer: OverlayAddress,
-        /// Reason for disconnection.
-        reason: Option<String>,
-    },
-}
-
-/// Events extracted from [`ClientEvent`] and routed to the pseudosettle service.
-#[derive(Debug, Clone)]
-pub enum PseudosettleEvent {
-    /// We sent a pseudosettle and received an ack.
-    Sent {
-        /// The peer we settled with.
-        peer: OverlayAddress,
-        /// The acknowledgment received.
-        ack: PaymentAck,
-    },
-    /// A peer sent us a pseudosettle request.
-    Received {
-        /// The peer that sent the request.
-        peer: OverlayAddress,
-        /// The payment amount requested.
-        amount: U256,
-        /// Request ID for sending ack.
-        request_id: u64,
-    },
-}
-
-/// Events extracted from [`ClientEvent`] and routed to the swap settlement service.
-#[cfg(feature = "swap")]
-#[derive(Debug, Clone)]
-pub enum SwapEvent {
-    /// We sent a cheque and the headers exchange completed.
-    ChequeSent {
-        /// The peer we sent the cheque to.
-        peer: OverlayAddress,
-        /// The peer's advertised exchange rate, from the headers exchange.
-        peer_rate: U256,
-    },
-    /// A peer sent us a cheque.
-    ChequeReceived {
-        /// The peer that sent the cheque.
-        peer: OverlayAddress,
-        /// The signed cheque received.
-        cheque: SignedCheque,
-        /// The peer's advertised exchange rate, from the headers exchange.
-        peer_rate: U256,
-    },
-}
+pub(crate) use vertex_swarm_client_protocol::{FailureKind, PushResponseTx, RetrievalResponseTx};
